v1-java 7 day

+++ newproject/src/challengeCh4/BankAccountTest.java
@@ -4,7 +4,7 @@ import java.util.Scanner;
 public class BankAccountTest {
   public static void main(String[] args) {
     boolean run = true;
-    int balance = 0;
+    long balance = 0;
     Scanner scanner = new Scanner(System.in);
     while(run) {
       System.out.println("-------------------------------------");
